[fxfs] Expand golden tests
Add a second unencrypted volume to go with the encrypted volume. There
are several flushes to ensure that some persistent layer files are used
in each. Many operations are left at the end after the flush so that
some will be replayed from the journal.

Bug: b/293937605
Change-Id: Ibb75f4edc9be18b363a27bfcfed5b589cf56a591
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024956
Reviewed-by: Brandon Castellano <bcastell@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>