[goldfish-control] Route the sysmem service to its outgoing directory.
The goldfish-control driver binds to a composite node
(goldfish-control-2) and gets the sysmem service from the "sysmem"
fragment node.

To provide its sysmem service to its descendents (for example,
goldfish-display), this change adds routing of the
fuchsia.hardware.sysmem/Service service to the device's outgoing
directory.

Note that this change has to add the sysmem device to the
"set_runtime_service_offers()" list, which is a hack so that the
node's binding property won't conflict with the existing sysmem
device.

On "ffx emu start":
Test: virtcon worked on core.x64
Test: "ffx driver list-devices -v", the goldfish-control node
      has correct bind properties and binds to goldfish-display.cm.
Test: "ffx driver list -v", goldfish-display.cm binds correctly.
Bug: 334988762
Bug: 334161084

Change-Id: Ib88ebec73c00fd9c3cf687b8d02b79ce16a1b0d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1028618
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>