[ffx][emu] Remove from being built-in
Removes ffx-emu from being built in to the main ffx binary, allowing it
to be run from the SDK instead.

This does not change how users invoke the tool, only how it executes
when `ffx emu` is invoked, but makes it so that both the tool and ffx
can build faster and with fewer resources making it easier to iterate on
them.

Bug: b/329326263
Change-Id: I7d98e0b09d768af06f4e40f33b89fe3898bf2612
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003651
Reviewed-by: Zijie He <zijiehe@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>