[software-compositor] Do manual comparisons rather than EXPECT_THAT(..., testing::ElementsAreArray(...))
For whatever reason, using `EXPECT_THAT(..., testing::ElementsAreArray(...))`
takes much longer to run compared to just doing manual comparisons via
`==`. On an uninstrumented bringup.arm64 build (without kvm) a loop for
a 1280x800 canvas would take ~30s with the current toolchain but less
than 1 second when just comparing individual container elements. The
perf difference happens to be more evident from the clang roll, so there
must be some upstream regression that led to it being even slower. I've
been able to assert that this isn't due to ubsan or hwasan. We also
don't see this on the asan builders because we just happen to not
timeout there for whatever reason.

While we investigate whatever upstream regression caused this, what we
can do in the meantime is just do individual element comparisons rather
than use EPXECT_THAT + testing::ElementsAreArray. This reduces a run on
the uninstrumented build to ~60s and a hwasan+ubsan instrumented build
to ~120s (on a qemu without kvm). This is much faster than the original
times we were seeing.

Bug: 333768776, 328267371
Change-Id: Ifa45cf12bdde43f39ce1d0f418f87713491b2f28
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1028371
Reviewed-by: Victor Costan <costan@google.com>
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>