[netstack3] Call egress hook before device layer
The call into the hgook was previously in the device layer because all
packets have to traverse there, whereas there were a few paths out of
the IP layer. This change moves the call to the egress hook as well as
the entry to the device layer into a helper function.

This change also ensures, through a witness type, that it's not possible
to call into the device layer without having called the egress hook.

Lastly, because this change moves MLD and IGMP to use a new method on
the IpLayerHandler trait, CoreCtxWithIpDeviceConfiguration required a
few new trait impls.

Bug: 328063820
Change-Id: I6c848f059398e88558e9e79316757da501851c4c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1025673
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>