[scrutiny] Add pre-signing golden file list
The scrutiny pre-signing verifier is configured to expect a dir path
containing goldens. However, if this path is provided as an input to
bazel, a warning about dependency checking of directories being unsound
will appear.

This change creates a list of all the goldens in the pre_signing dir
using glob, then forwards that list as an input to the scrutiny shell
action.

Bug: b/303923440
Test: fx build with https://fxrev.dev/1025377 locally
Change-Id: I7d26608c7aedad4ec27f148c6dc6ec2aec377eff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1029395
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Eric Stone <ecstone@google.com>