[kernel] Improve immediate access marking
Splits the MarkAccessed method into having an early before the lock
short circuit that is inlineable, and then the longer locked remainder.
This short circuit obviates the need for the reclamation checks in
UpdateOnAccessLocked, simplifying the logic.

The pf_flags passed in to UpdateOnAccessLocked had become redundant, as
LookupCursor is already skipping the call in the case of hardware page
faults.

Change-Id: I3c52b4d46370664616e90551157b4ab528942e22
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1026252
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>