[idk] Make idk rule hermetic.
This involved a slightly annoying amount of refactoring of the merge
process so that all the possible input files were made available to the
toplevel script, but I actually like it better this way.

Confirmed that it's a null-diff in the generated output (except for
helpdoc_out.tar.gz, which always seems to change).

Fixed: 333054435

Change-Id: I5d9a814e65cbdb1b73bc70577ca7120327c30fbf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1028373
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Turner <digit@google.com>
Fuchsia-Auto-Submit: Hunter Freyer <hjfreyer@google.com>