[assembly][networking] Add a http_client_stop_on_idle_timeout network config
This value turns into the `fuchsia.http-client.StopOnIdleTimeoutMillis`
configuration capability which is then routed to `http-client` to
control how it should stop on idle.

At this point no product sets this config so the configuration
capability will be missing, and http-client will detect that and
fallback to not stopping.

TEST: all the in-tree products assemble and boot.
TEST: http-client fall back to its default config correctly, from
  `ffx component show core/network/http-client`

Bug: b/333082616
Change-Id: Id5da2dfa658863d82ea6b402f0261e3f3aae741f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1029398
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>