[cmp-mgr] Use DebuglogResource in read-only logs
As part of the work to remove root resource, make use
of debuglog resource. This CL adds the routing to expose
the resource, uses the resource to make a read-only log
and updates the underlying syscall `debuglog_create` to
validate against the finer grained resource. Documentation
is updated where appropriate.

Bug: 31358
Change-Id: Id581564e067035027541828c8539852cc2d306e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1020196
Commit-Queue: Olivia Nordquist <nolivia@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Laura Peskin <pesk@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>