[ld] Introduce MockLoaderServiceForTest
This CL moves the logic of controlling the MockLoaderService from
LdLoadZirconLdsvcTestsBase to the new class MockLoaderServiceForTest.

The LdLoadZirconLdsvcTestsBase API becomes thin wrappers around the
MockLoaderServiceForTest API and is entirely defined in the *.h file to
hold the trivial code.

MockLoaderServiceForTest implements the business logic that was formerly
in LdLoadZirconLDsvcTestsBase, accepting a `GetVmo` template parameter
to specify the callback function that will return the VMO result for a
given `name`. It will lazily initialize a MockLoaderService. This API
can be shared with libdl and libld.

Bug: 324136435
Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs
Change-Id: I56f13ce840539ef2db2928ff7b268a0382bf8f78
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1028354
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>