[kernel][vm] Explicit accessed interval
Adds kernel cmdline option for explicitly controlling the accessed scan
interval instead of forcing it to be related to the min aging interval.
This is intended to allow for lowering the min aging interval without
increasing the mandatory accessed scan rate.

There is presently no control in the system assembly for changing the
min-aging-interval, so by defining the accessed-scan-interval default
value to be what the code would previously calculate there is not
behavioral change to any products.

Change-Id: I2ce94ddd5926f94ba8ea520dda389f4c0bfad2ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1025552
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: David Worsham <dworsham@google.com>