[build][info] Fix typo in Jiri hook
The script used an obsolete `--force-git` flag, which is now
the default, and disabled in the build actions with the flag
`--use-jiri-values`, but must not be used in the hook.

This should allow tqr/832170 to work correctly.

+ Move the .gitignore section to //build/info/jiri_generated/.gitignore
  for convenience, and to avoid the confusing pattern requirements
  in this file for paths relative to the root.

Bug: 335391299
Change-Id: I4c61a50bb3463042d0259a21c3a2fca0b366e090
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030874
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Turner <digit@google.com>