[mock-mmio-reg] Fold the -gtest and -zxtest rules together
This adds a new :mock-mmio-reg-zxtest target to the mock-mmio-reg
library. This target is functionally equivalent to the lib found at
//src/devices/testing/mock-mmio-reg-zxtest. Rather than copy the code,
we can just tuck the differences (which really boils down to whether
zxtest or gtest is used for the ASSERT_X macros) behind a -DFLAG build
define and #ifdef block.

The plan is to move all users back to this (new) target and only have
one version of the source to maintain. Once everyone is migrated back to
this target, the code at //s/d/t/mock-mmio-reg-zxtest can be deleted.

Bug: b/328812587
Change-Id: Ib8c8c8514c0e8cf457d511191f2b173c28710a30
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1029799
Commit-Queue: Sam Hansen <hansens@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>