[netstack3] Avoid dynamic dispatch and alloc
This has been bothering me for a while, it's rather trivial to implement
this iterator for device addresses and it saves us from a box and
dynamic dispatch.

This is still a problem in Slaac, but this is one unneeded dynamic
dispatch down.

Bug: 42059236
Change-Id: I50e714e58663ea1cb2add513e3172297dda61883
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031157
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>