[netstack3] Include routine/rule counts in inspect
Currently, when a hook has no routines or a routine has no rules, the
inspect data published by Netstack3 shows it as an empty node. This
makes adding an inspect selector for that node difficult without also
blanket selecting all properties contained by the node. Add properties
to the inspect data that indicate the number of routines installed on a
given hook and the number of rules in a given routine, so that there is
a property we can select on in privacy filters for an otherwise empty
inspect node.

Bug: 332702047
Test: netstack3-inspect-integration-test
Change-Id: I86041cd149dca6f09aa27708bd9d0b316e33967e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031314
Reviewed-by: Peter DeLong <pdelong@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>