[starnix][selinux] Create a clearer layout for the LSM/SELinux impl
The kernel's "selinux" subdirectory is renamed "security" to reflect
that the hooks and data structures it provides are in principle generic,
not tied to a specific choice of Linux Security Module.

Within the "security" directory a subdirectory holds the SELinux-specific
integration for the various hooks; since SELinux is the only LSM built
into Starnix some aspects are currently (re)published from the "security"
module.

The SELinux and procattr filesystems remain in the "security" directory
for now; the latter will be migrated to interface to SELinux via hooks,
similar to other parts of the kernel, and moved into the kernel's "fs"
module.

Bug: 335397745
Change-Id: Ia6ac8d0d794da5c2aed56d8f1b06d49235ba0d32
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030212
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Wez <wez@google.com>