[wlan][policy] Removes WLAN power control.
The low power listener is not used in practice and the
PHY APIs that the current implementation relies on are
going away.  This change removes the linkage to the power
client level service and the ability for wlancfg to set
power mode on PHYs.

Test: Flash DUT and ensure that it boots.
Bug: b/332904477
Change-Id: Ie305ff914cbe8c7140e9a0a23d19375a6e5c6f14
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030824
Reviewed-by: Marc Khouri <mnck@google.com>
Commit-Queue: Kevin Sakuma <sakuma@google.com>