[boot-shim][devicetree] fix redistributor-stride for 64-bit values
arm,gic-v3 redistributor-stride is represented sometimes as a single
cell property (32-bit) and sometimes as a two cell property (64-bit).
This change adds support to parse either and to export the maximal
mmio address range to the observer based on the region count, if used.

Bug: 330209521
Change-Id: I6cec89553650b4c1fdbf738fcb149181659e8316
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003694
Commit-Queue: Will Drewry <drewry@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>