[audio] Log device removals/errors
Log when a device has an error, and when it is removed,
in a consistent way to how we log when it is successfully
detected/queried/initialized/added.

Also set the default log settings as desired for integration
of a new non-DevFS Codec driver from the Bluetooth team.

Test: audio_services_device_registry_unittests
Test: audio_services_device_registry_warning_unittests

Bug: 321292607

Change-Id: I99830badfdf3edf0eabd6768370a9c78693a913b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030822
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>